2-get_bit.c

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -10,7 +10,7 @@ int get_bit(unsigned long int n, unsigned int index)
 {
 	int value;
 
-	value = (n >> index) & 1;
+	value = (int)(n >> index) & 1ul;
 
 	return (value);
 }
